pyproject.toml
[project]
name = "solvis-ha-integration"
version = "1.0.0"
description = "Home Assistant custom integration for Solvis heating controllers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-homeassistant-custom-component>=0.13",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
